third_party/blink/web_tests/external/wpt/resource-timing/entry-attributes.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing: PerformanceResourceTiming attributes</title>
<link rel="help" href="https://www.w3.org/TR/resource-timing-2/#sec-performanceresourcetiming"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/resource-loaders.js"></script>
<script src="resources/entry-invariants.js"></script>
<script>

// Given a resource-loader and a PerformanceResourceTiming validator, loads a
// resource and validates the resulting entry.
const attribute_test = (load_resource, validate, test_label) => {
  promise_test(
    async () => {
      // Clear out everything that isn't the one ResourceTiming entry under test.
      performance.clearResourceTimings();

      await load_resource();

      const entry_list = performance.getEntriesByType("resource");
      if (entry_list.length != 1) {
        throw new Error(`There should be one entry for one resource (found ${entry_list.length})`);
      }

      validate(entry_list[0]);
  }, test_label);
}

attribute_test(
  () => load.image("resources/fake_responses.py#hash=1"),
  entry => {
    assert_true(entry.name.includes('#hash=1'),
      "There should be a hash in the resource name");
    invariants.assert_http_resource(entry);
  },
  "Image resources should generate conformant entries");

attribute_test(
  () => load.font("/fonts/Ahem.ttf"),
  invariants.assert_http_resource,
  "Font resources should generate conformant entries");

attribute_test(
  () => load.image("/common/redirect.py?location=resources/fake_responses.py"),
  invariants.assert_same_origin_redirected_resource,
  "Same-origin redirects should populate redirectStart/redirectEnd");
</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that PerformanceResourceTiming entries' attributes are
populated with the correct values.</p>
</body>
</html>
